study_auth_module

+++ django/RECAP/articles/views.py
@@ -11,9 +11,14 @@ from django.contrib.auth.decorators import login_required
 # Create your views here.
 
 def index(request):
+    # embed()
+    visits_num = request.session.get('visits', 0)
+    request.session['visits'] = visits_num + 1
+    request.session.modified = True
     articles = Article.objects.all()
     context = {
         'articles': articles,
+        'visits': visits_num,
     }
     return render(request, 'articles/index.html', context=context)
 
@@ -29,7 +34,9 @@ def create(request):
             # title = form.cleaned_data.get('title')
             # content = form.cleaned_data.get('content')
             # article = Article.objects.create(title=title, content=content)
-            article = form.save()
+            article = form.save(commit=False)
+            article.user = request.user
+            article.save()
             return redirect(article)
         else:
             return redirect('articles:create')
@@ -45,19 +52,25 @@ def detail(request, article_pk, comment_pk=None):
     article = get_object_or_404(Article, pk=article_pk)
     comments = article.comment_set.all()
     commentform = CommentForm()
+    # visits_num = request.session.get('visits', 0)
+    # request.session['visits'] = visits_num + 1
+    # request.session.modified = True
     context = {
         'article': article,
         'comments': comments,
         'commentform': commentform,
-        'comment_pk': comment_pk
+        'comment_pk': comment_pk,
     }
     if comment_pk:
         comment = get_object_or_404(Comment, pk=comment_pk)
+        if comment.user != request.user():
+            return redirect(article)
         if request.method == 'POST':
             form = CommentForm(request.POST, instance=comment)
             if form.is_valid():
                 new_form = form.save(commit=False)
                 new_form.article = article
+                new_form.user = request.user
                 new_form.save()
                 return redirect(article)
         else:
@@ -70,6 +83,10 @@ def detail(request, article_pk, comment_pk=None):
 @login_required
 def update(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
+
+    if article.user != request.user:
+        return redirect(article)
+
     if request.method == 'POST':
         form = ArticleForm(request.POST, instance=article)
         
@@ -99,8 +116,13 @@ def update(request, article_pk):
 @require_POST
 @login_required
 def delete(request, article_pk):
+
+
     if request.user.is_authenticated:
         article = get_object_or_404(Article, pk=article_pk)
+        if article.user != request.user:
+            return redirect(article)
+
         if request.method == 'POST':
             article.delete()
             return redirect('articles:index')
@@ -119,6 +141,7 @@ def new_comment(request, article_pk):
         if form.is_valid():
             new_form = form.save(commit=False)
             new_form.article = Article.objects.get(pk=article_pk)
+            new_form.user = request.user
             new_form.save()
             return redirect(article)
     else:
@@ -133,4 +156,24 @@ def delete_comment(request, article_pk, comment_pk):
     if request.method == 'POST':
         if comment.article == article:
             comment.delete()
+    return redirect(article)
+
+
+def send_cookie(request):
+    res = HttpResponse('과자 받아라')
+    res.set_cookie('mycookie', 'oreo')
+    return res
+
+
+@login_required
+def like(request, article_pk):
+    article = get_object_or_404(Article,pk=article_pk)
+    article.like_users.add(request.user)
+    return redirect(article)
+
+
+@login_required
+def dislike(request, article_pk):
+    article = get_object_or_404(Article,pk=article_pk)
+    article.like_users.remove(request.user)
     return redirect(article)
